Setup TypeScript configs
