+++ package.json
@@ -4,19 +4,26 @@
   "description": "online tetris multiplayer",
   "main": "app.js",
   "scripts": {
-    "build": "webpack --config webpack-config.js",
-    "start": "node app.js",
-    "launch": "webpack --config webpack-config.js && node app.js"
+    "build": "npx tsc --project ./ts-config.json && webpack --config webpack-config.js",
+    "start": "node dist/typescript/app.js",
+    "launch": "npm run build && npm run start"
   },
   "author": "Avery",
   "license": "ISC",
   "dependencies": {
+    "@babel/core": "^7.23.7",
+    "@babel/preset-env": "^7.23.8",
+    "@babel/preset-react": "^7.23.3",
+    "@types/express": "^4.17.21",
+    "@types/pug": "^2.0.10",
+    "babel-loader": "^9.1.3",
     "css-loader": "^6.8.1",
     "express": "^4.18.2",
     "pug": "^3.0.2",
     "socket.io": "^4.7.2",
     "socket.io-client": "^4.7.2",
     "style-loader": "^3.3.3",
+    "typescript": "^5.3.3",
     "webpack": "^5.89.0",
     "webpack-cli": "^5.1.4"
   }
