+++ ts-config.json
@@ -1,0 +1,12 @@
+{
+    "compilerOptions": {
+        "module": "nodenext",
+        "noImplicitAny": true,
+        "removeComments": true,
+        "preserveConstEnums": true,
+        "sourceMap": true,
+        "allowJs": true,
+        "outDir": "./dist/typescript"
+    },
+    "include": ["./app.js", "./lib/**/*"]
+}
